Add files via upload
This code uses the elapsedMillis library to manage the time without using the traditional millis() function. The control LED function is responsible for toggling the LED state based on the specified interval. The state machine approach is used to determine when to toggle the LED based on the temperature conditions.